pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "onc-hydrophone-data"
version = "0.2.0"
description = "Tools for downloading and processing Ocean Networks Canada hydrophone data"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Spencer Bialek", email = "sbialek@uvic.ca"}
]
requires-python = ">=3.9"
dependencies = [
    "onc>=2.3.0",
    "numpy>=1.20.0",
    "scipy>=1.7.3",
    "matplotlib>=3.7.0",
    "pillow>=9.0.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "soundfile>=0.12.0",
    "librosa>=0.10.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
audio = ["torch>=2.0.0", "torchaudio>=2.0.0"]

[project.urls]
Homepage = "https://github.com/Spiffical/onc-hydrophone-data"
Repository = "https://github.com/Spiffical/onc-hydrophone-data"

[tool.setuptools.packages.find]
where = ["."]
include = ["onc_hydrophone_data*"]
